+++ services/venice-controller/src/main/java/com/linkedin/venice/controller/VeniceHelixAdmin.java
@@ -28,8 +28,6 @@ import static com.linkedin.venice.meta.VersionStatus.ROLLED_BACK;
 import static com.linkedin.venice.meta.VersionStatus.STARTED;
 import static com.linkedin.venice.pushmonitor.OfflinePushStatus.HELIX_ASSIGNMENT_COMPLETED;
 import static com.linkedin.venice.serialization.avro.AvroProtocolDefinition.PARTICIPANT_MESSAGE_SYSTEM_STORE_VALUE;
-import static com.linkedin.venice.system.store.MetaStoreWriter.KEY_STRING_STORE_NAME;
-import static com.linkedin.venice.utils.AvroSchemaUtils.isValidAvroSchema;
 import static com.linkedin.venice.utils.RegionUtils.isRegionPartOfRegionsFilterList;
 import static com.linkedin.venice.utils.RegionUtils.parseRegionsFilterList;
 import static com.linkedin.venice.views.ViewUtils.ETERNAL_TOPIC_RETENTION_ENABLED;
@@ -37,9 +35,6 @@ import static com.linkedin.venice.views.ViewUtils.LOG_COMPACTION_ENABLED;
 import static com.linkedin.venice.views.ViewUtils.PARTITION_COUNT;
 import static com.linkedin.venice.views.ViewUtils.USE_FAST_KAFKA_OPERATION_TIMEOUT;
 
-import com.linkedin.avroutil1.compatibility.AvroIncompatibleSchemaException;
-import com.linkedin.avroutil1.compatibility.RandomRecordGenerator;
-import com.linkedin.avroutil1.compatibility.RecordGenerationConfig;
 import com.linkedin.d2.balancer.D2Client;
 import com.linkedin.venice.ConfigKeys;
 import com.linkedin.venice.D2.D2ClientUtils;
@@ -103,7 +98,6 @@ import com.linkedin.venice.controllerapi.UpdateStoragePersonaQueryParams;
 import com.linkedin.venice.controllerapi.UpdateStoreQueryParams;
 import com.linkedin.venice.controllerapi.VersionResponse;
 import com.linkedin.venice.exceptions.ErrorType;
-import com.linkedin.venice.exceptions.InvalidVeniceSchemaException;
 import com.linkedin.venice.exceptions.ResourceStillExistsException;
 import com.linkedin.venice.exceptions.VeniceException;
 import com.linkedin.venice.exceptions.VeniceHttpException;
@@ -228,9 +222,6 @@ import com.linkedin.venice.schema.writecompute.DerivedSchemaEntry;
 import com.linkedin.venice.security.SSLFactory;
 import com.linkedin.venice.serialization.avro.AvroProtocolDefinition;
 import com.linkedin.venice.serialization.avro.InternalAvroSpecificSerializer;
-import com.linkedin.venice.serializer.AvroSerializer;
-import com.linkedin.venice.serializer.RecordDeserializer;
-import com.linkedin.venice.serializer.SerializerDeserializerFactory;
 import com.linkedin.venice.service.ICProvider;
 import com.linkedin.venice.stats.AbstractVeniceAggStats;
 import com.linkedin.venice.stats.ZkClientStatusStats;
@@ -242,7 +233,6 @@ import com.linkedin.venice.status.protocol.BatchJobHeartbeatValue;
 import com.linkedin.venice.status.protocol.PushJobDetails;
 import com.linkedin.venice.status.protocol.PushJobDetailsStatusTuple;
 import com.linkedin.venice.status.protocol.PushJobStatusRecordKey;
-import com.linkedin.venice.system.store.MetaStoreDataType;
 import com.linkedin.venice.system.store.MetaStoreReader;
 import com.linkedin.venice.system.store.MetaStoreWriter;
 import com.linkedin.venice.systemstore.schemas.StoreMetaKey;
@@ -359,7 +349,6 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       RedundantExceptionFilter.getRedundantExceptionFilter();
 
   private static final Logger LOGGER = LogManager.getLogger(VeniceHelixAdmin.class);
-  private static final int RECORD_COUNT = 10;
   public static final List<Class<? extends Throwable>> RETRY_FAILURE_TYPES = Collections.singletonList(Exception.class);
 
   private final VeniceControllerMultiClusterConfig multiClusterConfigs;
@@ -414,6 +403,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
   private final Lazy<PushStatusStoreWriter> pushStatusStoreWriter;
   private final SharedHelixReadOnlyZKSharedSystemStoreRepository zkSharedSystemStoreRepository;
   private final SharedHelixReadOnlyZKSharedSchemaRepository zkSharedSchemaRepository;
+  private final StoreSchemaService storeSchemaService;
   private final MetaStoreWriter metaStoreWriter;
   private final MetaStoreReader metaStoreReader;
   private final D2Client d2Client;
@@ -647,6 +637,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
         commonConfig.getSystemSchemaClusterName(),
         commonConfig.getRefreshAttemptsForZkReconnect(),
         commonConfig.getRefreshIntervalForZkReconnectInMs());
+    storeSchemaService = new StoreSchemaService(this);
     metaStoreWriter = new MetaStoreWriter(
         topicManagerRepository.getLocalTopicManager(),
         veniceWriterFactory,
@@ -3862,14 +3853,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       String storeName,
       int valueSchemaID,
       int rmdVersionID) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepo = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    SchemaEntry schemaEntry = schemaRepo.getReplicationMetadataSchema(storeName, valueSchemaID, rmdVersionID);
-    if (schemaEntry == null) {
-      return Optional.empty();
-    } else {
-      return Optional.of(schemaEntry.getSchema());
-    }
+    return storeSchemaService.getReplicationMetadataSchema(clusterName, storeName, valueSchemaID, rmdVersionID);
   }
 
   @Override
@@ -5380,45 +5364,11 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
 
   @Override
   public Set<Integer> getInUseValueSchemaIds(String clusterName, String storeName) {
-    if (isParent()) {
-      return Collections.emptySet();
-    }
-
-    Store store = getStore(clusterName, storeName);
-    Set<Integer> schemaIds = new HashSet<>();
-
-    // Fetch value schema id used by all existing store version
-    for (Version version: store.getVersions()) {
-      Map<String, String> map = new HashMap<>(2);
-      map.put(KEY_STRING_STORE_NAME, storeName);
-      map.put(MetaStoreWriter.KEY_STRING_VERSION_NUMBER, Integer.toString(version.getNumber()));
-      StoreMetaKey key = MetaStoreDataType.VALUE_SCHEMAS_WRITTEN_PER_STORE_VERSION.getStoreMetaKey(map);
-      StoreMetaValue metaValue = getMetaStoreValue(key, storeName);
-
-      if (metaValue == null) {
-        String msg = "Could not find in-use value schema for store " + storeName;
-        LOGGER.warn(msg);
-        throw new VeniceException(msg);
-      }
-      schemaIds.addAll(metaValue.storeValueSchemaIdsWrittenPerStoreVersion);
-    }
-    return schemaIds;
+    return storeSchemaService.getInUseValueSchemaIds(clusterName, storeName);
   }
 
   public void deleteValueSchemas(String clusterName, String storeName, Set<Integer> unusedValueSchemaIds) {
-    Set<Integer> inuseValueSchemaIds = getInUseValueSchemaIds(clusterName, storeName);
-    boolean isCommon = unusedValueSchemaIds.stream().anyMatch(inuseValueSchemaIds::contains);
-    if (isCommon) {
-      String msg = "For store " + storeName + " cannot delete value schema ids they being used. schema ids: "
-          + unusedValueSchemaIds;
-      LOGGER.error(msg);
-      throw new VeniceException(msg);
-    }
-    // delete the unused schemas.
-    unusedValueSchemaIds.forEach(id -> {
-      getHelixVeniceClusterResources(clusterName).getSchemaRepository().removeValueSchema(storeName, id);
-      LOGGER.info("Removed value schema with ID " + id + " for store " + storeName);
-    });
+    storeSchemaService.deleteValueSchemas(clusterName, storeName, unusedValueSchemaIds);
   }
 
   private void setStoreCompressionStrategy(
@@ -6962,9 +6912,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
    */
   @Override
   public SchemaEntry getKeySchema(String clusterName, String storeName) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepo = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    return schemaRepo.getKeySchema(storeName);
+    return storeSchemaService.getKeySchema(clusterName, storeName);
   }
 
   /**
@@ -6972,9 +6920,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
    */
   @Override
   public Collection<SchemaEntry> getValueSchemas(String clusterName, String storeName) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepo = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    return schemaRepo.getValueSchemas(storeName);
+    return storeSchemaService.getValueSchemas(clusterName, storeName);
   }
 
   /**
@@ -6982,9 +6928,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
    */
   @Override
   public Collection<DerivedSchemaEntry> getDerivedSchemas(String clusterName, String storeName) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepo = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    return schemaRepo.getDerivedSchemas(storeName);
+    return storeSchemaService.getDerivedSchemas(clusterName, storeName);
   }
 
   /**
@@ -6992,15 +6936,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
    */
   @Override
   public int getValueSchemaId(String clusterName, String storeName, String valueSchemaStr) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepo = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    int schemaId = schemaRepo.getValueSchemaId(storeName, valueSchemaStr);
-    // validate the schema as VPJ uses this method to fetch the value schema. Fail loudly if the schema user trying
-    // to push is bad.
-    if (schemaId != SchemaData.INVALID_VALUE_SCHEMA_ID) {
-      AvroSchemaUtils.validateAvroSchemaStr(valueSchemaStr);
-    }
-    return schemaId;
+    return storeSchemaService.getValueSchemaId(clusterName, storeName, valueSchemaStr);
   }
 
   /**
@@ -7008,15 +6944,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
    */
   @Override
   public GeneratedSchemaID getDerivedSchemaId(String clusterName, String storeName, String schemaStr) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepo = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    GeneratedSchemaID schemaID = schemaRepo.getDerivedSchemaId(storeName, schemaStr);
-    // validate the schema as VPJ uses this method to fetch the value schema. Fail loudly if the schema user trying
-    // to push is bad.
-    if (schemaID.isValid()) {
-      AvroSchemaUtils.validateAvroSchemaStr(schemaStr);
-    }
-    return schemaID;
+    return storeSchemaService.getDerivedSchemaId(clusterName, storeName, schemaStr);
   }
 
   /**
@@ -7024,82 +6952,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
    */
   @Override
   public SchemaEntry getValueSchema(String clusterName, String storeName, int id) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepo = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    return schemaRepo.getValueSchema(storeName, id);
-  }
-
-  private void validateValueSchemaUsingRandomGenerator(String schemaStr, String clusterName, String storeName) {
-    VeniceControllerClusterConfig config = getHelixVeniceClusterResources(clusterName).getConfig();
-    if (!config.isControllerSchemaValidationEnabled()) {
-      return;
-    }
-
-    ReadWriteSchemaRepository schemaRepository = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    Collection<SchemaEntry> schemaEntries = schemaRepository.getValueSchemas(storeName);
-    AvroSerializer serializer;
-    Schema existingSchema = null;
-
-    Schema newSchema = AvroSchemaParseUtils.parseSchemaFromJSONStrictValidation(schemaStr);
-    RandomRecordGenerator recordGenerator = new RandomRecordGenerator();
-    RecordGenerationConfig genConfig = RecordGenerationConfig.newConfig().withAvoidNulls(true);
-
-    for (int i = 0; i < RECORD_COUNT; i++) {
-      // check if new records written with new schema can be read using existing older schema
-      // Object record =
-      Object record = recordGenerator.randomGeneric(newSchema, genConfig);
-      serializer = new AvroSerializer<>(newSchema);
-      byte[] bytes = serializer.serialize(record);
-      for (SchemaEntry schemaEntry: schemaEntries) {
-        try {
-          existingSchema = schemaEntry.getSchema();
-          if (!isValidAvroSchema(existingSchema)) {
-            LOGGER.warn("Skip validating ill-formed schema for store: {}", storeName);
-            continue;
-          }
-          RecordDeserializer<Object> deserializer =
-              SerializerDeserializerFactory.getAvroGenericDeserializer(newSchema, existingSchema);
-          deserializer.deserialize(bytes);
-        } catch (Exception e) {
-          if (e instanceof AvroIncompatibleSchemaException) {
-            LOGGER.warn("Found incompatible avro schema with bad union branch for store: {}", storeName, e);
-            continue;
-          }
-          throw new InvalidVeniceSchemaException(
-              "Error while trying to add new schema: " + schemaStr + "  for store " + storeName
-                  + " as it is incompatible with existing schema: " + existingSchema,
-              e);
-        }
-      }
-    }
-
-    // check if records written with older schema can be read using the new schema
-    for (int i = 0; i < RECORD_COUNT; i++) {
-      for (SchemaEntry schemaEntry: schemaEntries) {
-        try {
-          Object record = recordGenerator.randomGeneric(schemaEntry.getSchema(), genConfig);
-          serializer = new AvroSerializer(schemaEntry.getSchema());
-          byte[] bytes = serializer.serialize(record);
-          existingSchema = schemaEntry.getSchema();
-          if (!isValidAvroSchema(existingSchema)) {
-            LOGGER.warn("Skip validating ill-formed schema for store: {}", storeName);
-            continue;
-          }
-          RecordDeserializer<Object> deserializer =
-              SerializerDeserializerFactory.getAvroGenericDeserializer(existingSchema, newSchema);
-          deserializer.deserialize(bytes);
-        } catch (Exception e) {
-          if (e instanceof AvroIncompatibleSchemaException) {
-            LOGGER.warn("Found incompatible avro schema with bad union branch for store: {}", storeName, e);
-            continue;
-          }
-          throw new InvalidVeniceSchemaException(
-              "Error while trying to add new schema: " + schemaStr + "  for store " + storeName
-                  + " as it is incompatible with existing schema: " + existingSchema,
-              e);
-        }
-      }
-    }
+    return storeSchemaService.getValueSchema(clusterName, storeName, id);
   }
 
   /**
@@ -7111,15 +6964,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       String storeName,
       String valueSchemaStr,
       DirectionalSchemaCompatibilityType expectedCompatibilityType) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepository = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    SchemaEntry schemaEntry = schemaRepository.addValueSchema(storeName, valueSchemaStr, expectedCompatibilityType);
-    // For duplicates, addValueSchema returns DUPLICATE_VALUE_SCHEMA_CODE; look up the real id so callers
-    // (e.g. SchemaRoutes) always receive a concrete schema id in the response.
-    int returnId = schemaEntry.getId() == SchemaData.DUPLICATE_VALUE_SCHEMA_CODE
-        ? schemaRepository.getValueSchemaId(storeName, valueSchemaStr)
-        : schemaEntry.getId();
-    return new SchemaEntry(returnId, valueSchemaStr);
+    return storeSchemaService.addValueSchema(clusterName, storeName, valueSchemaStr, expectedCompatibilityType);
   }
 
   /**
@@ -7134,18 +6979,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       String valueSchemaStr,
       int schemaId,
       DirectionalSchemaCompatibilityType compatibilityType) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepository = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    int newValueSchemaId =
-        schemaRepository.preCheckValueSchemaAndGetNextAvailableId(storeName, valueSchemaStr, compatibilityType);
-    if (newValueSchemaId != SchemaData.DUPLICATE_VALUE_SCHEMA_CODE && newValueSchemaId != schemaId) {
-      throw new VeniceException(
-          "Inconsistent value schema id between the caller and the local schema repository."
-              + " Expected new schema id of " + schemaId + " but the next available id from the local repository is "
-              + newValueSchemaId + " for store " + storeName + " in cluster " + clusterName + " Schema: "
-              + valueSchemaStr);
-    }
-    return schemaRepository.addValueSchema(storeName, valueSchemaStr, newValueSchemaId);
+    return storeSchemaService.addValueSchema(clusterName, storeName, valueSchemaStr, schemaId, compatibilityType);
   }
 
   /**
@@ -7159,14 +6993,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       String storeName,
       int valueSchemaId,
       String derivedSchemaStr) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepository = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    schemaRepository.addDerivedSchema(storeName, derivedSchemaStr, valueSchemaId);
-
-    return new DerivedSchemaEntry(
-        valueSchemaId,
-        schemaRepository.getDerivedSchemaId(storeName, derivedSchemaStr).getGeneratedSchemaVersion(),
-        derivedSchemaStr);
+    return storeSchemaService.addDerivedSchema(clusterName, storeName, valueSchemaId, derivedSchemaStr);
   }
 
   /**
@@ -7180,9 +7007,8 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       int valueSchemaId,
       int derivedSchemaId,
       String derivedSchemaStr) {
-    checkControllerLeadershipFor(clusterName);
-    return getHelixVeniceClusterResources(clusterName).getSchemaRepository()
-        .addDerivedSchema(storeName, derivedSchemaStr, valueSchemaId, derivedSchemaId);
+    return storeSchemaService
+        .addDerivedSchema(clusterName, storeName, valueSchemaId, derivedSchemaId, derivedSchemaStr);
   }
 
   /**
@@ -7194,9 +7020,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       String storeName,
       int valueSchemaId,
       int derivedSchemaId) {
-    checkControllerLeadershipFor(clusterName);
-    return getHelixVeniceClusterResources(clusterName).getSchemaRepository()
-        .removeDerivedSchema(storeName, valueSchemaId, derivedSchemaId);
+    return storeSchemaService.removeDerivedSchema(clusterName, storeName, valueSchemaId, derivedSchemaId);
   }
 
   /**
@@ -7213,26 +7037,8 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       int valueSchemaId,
       String supersetSchemaStr,
       int supersetSchemaId) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepository = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-
-    final SchemaEntry existingSupersetSchemaEntry = schemaRepository.getValueSchema(storeName, supersetSchemaId);
-    if (existingSupersetSchemaEntry == null) {
-      // If the new superset schema does not exist in the schema repo, add it
-      LOGGER.info("Adding superset schema: {} for store: {}", supersetSchemaStr, storeName);
-      schemaRepository.addValueSchema(storeName, supersetSchemaStr, supersetSchemaId);
-
-    } else {
-      final Schema newSupersetSchema = AvroSchemaParseUtils.parseSchemaFromJSONStrictValidation(supersetSchemaStr);
-      if (!AvroSchemaUtils.compareSchemaIgnoreFieldOrder(existingSupersetSchemaEntry.getSchema(), newSupersetSchema)) {
-        throw new VeniceException(
-            "Existing schema with id " + existingSupersetSchemaEntry.getId() + " does not match with new schema "
-                + supersetSchemaStr);
-      }
-    }
-
-    // add the value schema
-    return schemaRepository.addValueSchema(storeName, valueSchema, valueSchemaId);
+    return storeSchemaService
+        .addSupersetSchema(clusterName, storeName, valueSchema, valueSchemaId, supersetSchemaStr, supersetSchemaId);
   }
 
   int getValueSchemaIdIgnoreFieldOrder(
@@ -7240,16 +7046,8 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       String storeName,
       String valueSchemaStr,
       Comparator<Schema> schemaComparator) {
-    checkControllerLeadershipFor(clusterName);
-    SchemaEntry valueSchemaEntry = new SchemaEntry(SchemaData.UNKNOWN_SCHEMA_ID, valueSchemaStr);
-
-    for (SchemaEntry schemaEntry: getValueSchemas(clusterName, storeName)) {
-      if (schemaComparator.compare(schemaEntry.getSchema(), valueSchemaEntry.getSchema()) == 0) {
-        return schemaEntry.getId();
-      }
-    }
-    return SchemaData.INVALID_VALUE_SCHEMA_ID;
-
+    return storeSchemaService
+        .getValueSchemaIdIgnoreFieldOrder(clusterName, storeName, valueSchemaStr, schemaComparator);
   }
 
   int checkPreConditionForAddValueSchemaAndGetNewSchemaId(
@@ -7257,12 +7055,11 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       String storeName,
       String valueSchemaStr,
       DirectionalSchemaCompatibilityType expectedCompatibilityType) {
-    AvroSchemaUtils.validateAvroSchemaStr(valueSchemaStr);
-    AvroSchemaParseUtils.parseSchemaFromJSONStrictValidation(valueSchemaStr);
-    validateValueSchemaUsingRandomGenerator(valueSchemaStr, clusterName, storeName);
-    checkControllerLeadershipFor(clusterName);
-    return getHelixVeniceClusterResources(clusterName).getSchemaRepository()
-        .preCheckValueSchemaAndGetNextAvailableId(storeName, valueSchemaStr, expectedCompatibilityType);
+    return storeSchemaService.checkPreConditionForAddValueSchemaAndGetNewSchemaId(
+        clusterName,
+        storeName,
+        valueSchemaStr,
+        expectedCompatibilityType);
   }
 
   int checkPreConditionForAddDerivedSchemaAndGetNewSchemaId(
@@ -7270,9 +7067,8 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       String storeName,
       int valueSchemaId,
       String derivedSchemaStr) {
-    checkControllerLeadershipFor(clusterName);
-    return getHelixVeniceClusterResources(clusterName).getSchemaRepository()
-        .preCheckDerivedSchemaAndGetNextAvailableId(storeName, valueSchemaId, derivedSchemaStr);
+    return storeSchemaService
+        .checkPreConditionForAddDerivedSchemaAndGetNewSchemaId(clusterName, storeName, valueSchemaId, derivedSchemaStr);
   }
 
   /**
@@ -7280,9 +7076,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
    */
   @Override
   public Collection<RmdSchemaEntry> getReplicationMetadataSchemas(String clusterName, String storeName) {
-    checkControllerLeadershipFor(clusterName);
-    ReadWriteSchemaRepository schemaRepo = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    return schemaRepo.getReplicationMetadataSchemas(storeName);
+    return storeSchemaService.getReplicationMetadataSchemas(clusterName, storeName);
   }
 
   boolean checkIfValueSchemaAlreadyHasRmdSchema(
@@ -7290,16 +7084,8 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       String storeName,
       final int valueSchemaID,
       final int replicationMetadataVersionId) {
-    checkControllerLeadershipFor(clusterName);
-    Collection<RmdSchemaEntry> schemaEntries =
-        getHelixVeniceClusterResources(clusterName).getSchemaRepository().getReplicationMetadataSchemas(storeName);
-    for (RmdSchemaEntry rmdSchemaEntry: schemaEntries) {
-      if (rmdSchemaEntry.getValueSchemaID() == valueSchemaID
-          && rmdSchemaEntry.getId() == replicationMetadataVersionId) {
-        return true;
-      }
-    }
-    return false;
+    return storeSchemaService
+        .checkIfValueSchemaAlreadyHasRmdSchema(clusterName, storeName, valueSchemaID, replicationMetadataVersionId);
   }
 
   boolean checkIfMetadataSchemaAlreadyPresent(
@@ -7307,19 +7093,8 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       String storeName,
       int valueSchemaId,
       RmdSchemaEntry rmdSchemaEntry) {
-    checkControllerLeadershipFor(clusterName);
-    try {
-      Collection<RmdSchemaEntry> schemaEntries =
-          getHelixVeniceClusterResources(clusterName).getSchemaRepository().getReplicationMetadataSchemas(storeName);
-      for (RmdSchemaEntry schemaEntry: schemaEntries) {
-        if (schemaEntry.equals(rmdSchemaEntry)) {
-          return true;
-        }
-      }
-    } catch (Exception e) {
-      LOGGER.warn("Exception in checkIfMetadataSchemaAlreadyPresent ", e);
-    }
-    return false;
+    return storeSchemaService
+        .checkIfMetadataSchemaAlreadyPresent(clusterName, storeName, valueSchemaId, rmdSchemaEntry);
   }
 
   /**
@@ -7334,28 +7109,12 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
       int valueSchemaId,
       int replicationMetadataVersionId,
       String replicationMetadataSchemaStr) {
-    checkControllerLeadershipFor(clusterName);
-
-    RmdSchemaEntry rmdSchemaEntry =
-        new RmdSchemaEntry(valueSchemaId, replicationMetadataVersionId, replicationMetadataSchemaStr);
-    if (checkIfMetadataSchemaAlreadyPresent(clusterName, storeName, valueSchemaId, rmdSchemaEntry)) {
-      LOGGER.info(
-          "Timestamp metadata schema Already present: for store: {} in cluster: {} metadataSchema: {} "
-              + "replicationMetadataVersionId: {} valueSchemaId: {}",
-          storeName,
-          clusterName,
-          replicationMetadataSchemaStr,
-          replicationMetadataVersionId,
-          valueSchemaId);
-      return rmdSchemaEntry;
-    }
-
-    return getHelixVeniceClusterResources(clusterName).getSchemaRepository()
-        .addReplicationMetadataSchema(
-            storeName,
-            valueSchemaId,
-            replicationMetadataSchemaStr,
-            replicationMetadataVersionId);
+    return storeSchemaService.addReplicationMetadataSchema(
+        clusterName,
+        storeName,
+        valueSchemaId,
+        replicationMetadataVersionId,
+        replicationMetadataSchemaStr);
   }
 
   /**
@@ -7469,10 +7228,7 @@ public class VeniceHelixAdmin implements Admin, StoreCleaner {
   }
 
   Schema getSupersetOrLatestValueSchema(String clusterName, Store store) {
-    ReadWriteSchemaRepository schemaRepository = getHelixVeniceClusterResources(clusterName).getSchemaRepository();
-    // If already a superset schema exists, try to generate the new superset from that and the input value schema
-    SchemaEntry existingSchema = schemaRepository.getSupersetOrLatestValueSchema(store.getName());
-    return existingSchema == null ? null : existingSchema.getSchema();
+    return storeSchemaService.getSupersetOrLatestValueSchema(clusterName, store);
   }
 
   /**
